+++ tfe/resource_tfe_no_code_module.go
@@ -55,23 +55,24 @@ func resourceTFENoCodeModule() *schema.Resource {
 			"variable_options": {
 				Type:     schema.TypeList,
 				Optional: true,
-				ForceNew: true,
+				ForceNew: false,
 				Elem: &schema.Resource{
 					Schema: map[string]*schema.Schema{
 						"name": {
 							Type:     schema.TypeString,
 							Required: true,
-							ForceNew: true,
+							ForceNew: false,
 						},
 						"type": {
 							Type:     schema.TypeString,
 							Required: true,
-							ForceNew: true,
+							ForceNew: false,
 						},
 						"options": {
 							Type:     schema.TypeList,
-							ForceNew: true,
-							Optional: true,
+							ForceNew: false,
+							Required: true,
+							MinItems: 1,
 							Elem:     &schema.Schema{Type: schema.TypeString},
 						},
 					},