+++ tfe/resource_tfe_registry_module.go
@@ -136,7 +136,12 @@ func resourceTFERegistryModuleCreateWithoutVCS(meta interface{}, d *schema.Resou
 	options := tfe.RegistryModuleCreateOptions{
 		Name:     tfe.String(d.Get("name").(string)),
 		Provider: tfe.String(d.Get("module_provider").(string)),
-		NoCode:   d.Get("no_code").(bool),
+	}
+
+	if v, ok := d.GetOk("no_code"); ok {
+		// TODO: update the deprecation message with the version that will remove this.
+		log.Println("[WARN] no_code is deprecated as of release 0.33.0 and may be removed in a future version. The preferred way to create a no-code registry module is to create a tfe_nocode_module resource for the tfe_registry_module resource.")
+		options.NoCode = tfe.Bool(v.(bool))
 	}
 
 	if registryName, ok := d.GetOk("registry_name"); ok {
@@ -212,9 +217,13 @@ func resourceTFERegistryModuleCreate(d *schema.ResourceData, meta interface{}) e
 func resourceTFERegistryModuleUpdate(d *schema.ResourceData, meta interface{}) error {
 	config := meta.(ConfiguredClient)
 
-	options := tfe.RegistryModuleUpdateOptions{
-		NoCode: tfe.Bool(d.Get("no_code").(bool)),
+	options := tfe.RegistryModuleUpdateOptions{}
+	if v, ok := d.GetOk("no_code"); ok {
+		// TODO: update the deprecation message with the version that will remove this.
+		log.Println("[WARN] no_code is deprecated as of release 0.33.0 and may be removed in a future version. The preferred way to create a no-code registry module is to create a tfe_nocode_module resource for the tfe_registry_module resource.")
+		options.NoCode = tfe.Bool(v.(bool))
 	}
+
 	var registryModule *tfe.RegistryModule
 	var err error